requirements.txt
Flask
requests
cryptography
watchdog
python-dotenv